+++ dlint/linters/bad_requests_use.py
@@ -27,42 +27,42 @@ class BadRequestsUseLinter(bad_kwarg_use.BadKwargUseLinter):
     def kwargs(self):
         return [
             {
-                "attribute_name": "request",
+                "module_path": "requests.request",
                 "kwarg_name": "verify",
                 "predicate": tree.kwarg_false,
             },
             {
-                "attribute_name": "get",
+                "module_path": "requests.get",
                 "kwarg_name": "verify",
                 "predicate": tree.kwarg_false,
             },
             {
-                "attribute_name": "options",
+                "module_path": "requests.options",
                 "kwarg_name": "verify",
                 "predicate": tree.kwarg_false,
             },
             {
-                "attribute_name": "head",
+                "module_path": "requests.head",
                 "kwarg_name": "verify",
                 "predicate": tree.kwarg_false,
             },
             {
-                "attribute_name": "post",
+                "module_path": "requests.post",
                 "kwarg_name": "verify",
                 "predicate": tree.kwarg_false,
             },
             {
-                "attribute_name": "put",
+                "module_path": "requests.put",
                 "kwarg_name": "verify",
                 "predicate": tree.kwarg_false,
             },
             {
-                "attribute_name": "patch",
+                "module_path": "requests.patch",
                 "kwarg_name": "verify",
                 "predicate": tree.kwarg_false,
             },
             {
-                "attribute_name": "delete",
+                "module_path": "requests.delete",
                 "kwarg_name": "verify",
                 "predicate": tree.kwarg_false,
             },