+++ dlint/linters/helpers/bad_kwarg_use.py
@@ -9,6 +9,7 @@ from __future__ import (
 
 import abc
 import ast
+import warnings
 
 from .. import base
 from ... import tree
@@ -28,7 +29,7 @@ class BadKwargUseLinter(base.BaseLinter, util.ABC):
 
             [
                 {
-                    "attribute_name": "name1|mod1.mod2.name1",
+                    "attribute_name": "name1", | "module_path": "mod1.mod2.name1",
                     "kwarg_name": "kwarg1",
                     "predicate": <function>,
                 },
@@ -38,14 +39,13 @@ class BadKwargUseLinter(base.BaseLinter, util.ABC):
         function that takes the Call object and a kwarg name and returns
         True|False.
 
-        The "attribute_name" has been overloaded to behave differently if a
-        period (".") is included or not. This key can now take a value like
-        "attribute" or "module1.module2.attribute". This is a temporary,
-        backwards-compatible change to allow us to more accurately specify an
-        attribute from a specific module. E.g. "subprocess.Popen(shell=True)"
-        vs. just "Popen(shell=True)". In the future we should deprecate and
-        eventually remove the old behavior in favor of fully specified module
-        attributes.
+        Either "attribute_name" or "module_path" can be specified, however
+        "attribute_name" is deprecated and will be removed in the future.
+        This is a temporary, backwards-compatible change to allow us to more
+        accurately specify an attribute from a specific module. E.g.
+        "subprocess.Popen(shell=True)" vs. just "Popen(shell=True)". In the
+        future we will remove the old behavior in favor of fully specified
+        module attributes.
         """
 
     def visit_Call(self, node):
@@ -54,20 +54,24 @@ class BadKwargUseLinter(base.BaseLinter, util.ABC):
         if not isinstance(node.func, (ast.Attribute, ast.Name)):
             return
 
-        def compare_attribute_name_and_call(attribute, call_node):
-            if "." not in attribute:
-                return attribute == tree.call_name(call_node)
+        def compare_kwarg_and_call(kwarg, call_node):
+            if "attribute_name" in kwarg:
+                warnings.warn(
+                    "'attribute_name' deprecated, please use fully specified 'module_path'",
+                    DeprecationWarning
+                )
+                return kwarg["attribute_name"] == tree.call_name(call_node)
 
             module_path = ".".join(tree.module_path(node.func))
 
             return self.namespace.illegal_module_imported(
                 module_path,
-                attribute
+                kwarg["module_path"]
             )
 
         bad_kwarg = any(
             (
-                compare_attribute_name_and_call(kwarg["attribute_name"], node)
+                compare_kwarg_and_call(kwarg, node)
                 and kwarg["predicate"](node, kwarg["kwarg_name"])
             )
             for kwarg in self.kwargs