+++ tests/test_helpers/test_bad_kwarg_use.py
@@ -11,6 +11,8 @@ import unittest
 
 import dlint
 
+import pytest
+
 
 def get_bad_kwarg_use_implementation(kwargs):
     class Cls(dlint.linters.helpers.bad_kwarg_use.BadKwargUseLinter):
@@ -26,6 +28,7 @@ def get_bad_kwarg_use_implementation(kwargs):
 
 class TestBadKwargUse(dlint.test.base.BaseTest):
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_empty(self):
         python_node = self.get_ast_node(
             """
@@ -48,6 +51,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_empty_kwargs(self):
         python_node = self.get_ast_node(
             """
@@ -63,6 +67,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_present_true(self):
         python_node = self.get_ast_node(
             """
@@ -94,6 +99,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_present_false(self):
         python_node = self.get_ast_node(
             """
@@ -119,6 +125,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_present_true_missing(self):
         python_node = self.get_ast_node(
             """
@@ -144,6 +151,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_present_false_missing(self):
         python_node = self.get_ast_node(
             """
@@ -175,6 +183,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_same_name_different_kwarg(self):
         python_node = self.get_ast_node(
             """
@@ -200,6 +209,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_different_name_same_kwarg(self):
         python_node = self.get_ast_node(
             """
@@ -225,6 +235,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_multi(self):
         python_node = self.get_ast_node(
             """
@@ -267,6 +278,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_nested_calls(self):
         python_node = self.get_ast_node(
             """
@@ -298,6 +310,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_subscript_calls(self):
         python_node = self.get_ast_node(
             """
@@ -329,6 +342,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_primitive_true(self):
         python_node = self.get_ast_node(
             """
@@ -360,6 +374,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_primitive_false(self):
         python_node = self.get_ast_node(
             """
@@ -391,6 +406,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    @pytest.mark.filterwarnings('ignore::DeprecationWarning:dlint')
     def test_kwargs_primitive_none(self):
         python_node = self.get_ast_node(
             """
@@ -434,7 +450,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
         linter = get_bad_kwarg_use_implementation(
             [
                 {
-                    "attribute_name": "foo.bar.baz",
+                    "module_path": "foo.bar.baz",
                     "kwarg_name": "kwarg",
                     "predicate": dlint.tree.kwarg_present,
                 },
@@ -466,7 +482,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
         linter = get_bad_kwarg_use_implementation(
             [
                 {
-                    "attribute_name": "foo.bar.baz",
+                    "module_path": "foo.bar.baz",
                     "kwarg_name": "kwarg",
                     "predicate": dlint.tree.kwarg_present,
                 },
@@ -491,7 +507,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
         linter = get_bad_kwarg_use_implementation(
             [
                 {
-                    "attribute_name": "foo.bar.baz",
+                    "module_path": "foo.bar.baz",
                     "kwarg_name": "kwarg",
                     "predicate": dlint.tree.kwarg_present,
                 },
@@ -522,7 +538,7 @@ class TestBadKwargUse(dlint.test.base.BaseTest):
         linter = get_bad_kwarg_use_implementation(
             [
                 {
-                    "attribute_name": "foo.bar.baz",
+                    "module_path": "foo.bar.baz",
                     "kwarg_name": "kwarg",
                     "predicate": dlint.tree.kwarg_not_present,
                 },